--- a/src/Norg.jl
+++ b/src/Norg.jl
@@ -24,18 +24,19 @@ include("tokenize.jl")
 include("ast.jl")
 include("match/match.jl")
 include("parser/parser.jl")
-# include("codegen.jl")
+include("codegen.jl")
 
 using .Kinds
 using .Tokens
 using .Scanners
 using .Tokenize
 using .AST
-# using .Parser
-# using .Codegen
-#
-# Base.parse(::Type{HTMLTarget}, s) = codegen(HTMLTarget, parse_norg(tokenize(s)))
-#
-# export HTMLTarget
+using .Parser
+using .Codegen
+
+Base.parse(::Type{HTMLTarget}, s) = codegen(HTMLTarget, parse_norg(tokenize(s)))
+Base.parse(t::HTMLTarget, s) = codegen(t, parse_norg(tokenize(s)))
+
+export HTMLTarget
 
 end