--- a/src/codegen.jl
+++ b/src/codegen.jl
@@ -1,10 +1,12 @@
 module Codegen
 using ..AST
+using ..Strategies
+using ..Kinds
 
 abstract type CodegenTarget end
 
 function codegen end
-codegen(t::Type{T}, ast) where {T <: CodegenTarget} = codegen(t(), ast)
+codegen(t::Type{T}, ast::AST.NorgDocument) where {T <: CodegenTarget} = codegen(t(), ast)
 
 include("codegen/html.jl")
 using .HTMLCodegen