Well, kinda. There are still some issues, but it's already much quicker than before.

Before type optimization :

```julia
julia> using BenchmarkTools, ProfileView, Debugger, Infiltrator, Revise, AbstractTrees, Test
[ Info: Precompiling BenchmarkTools [6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf]

julia> using Norg

julia> s = """
       [Neorg] is a fancy organizational tool for everyone.

       I like shilling [Neorg]{https://github.com/nvim-neorg/neorg} sorry :(
       """
"[Neorg] is a fancy organizational tool for everyone.\n\nI like shilling [Neorg]{https://github.com/nvim-neorg/neorg} sorry :(\n"

julia> @benchmark parse(Norg.AST.NorgDocument, s)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  352.846 μs …   4.289 ms  ┊ GC (min … max): 0.00% … 86.13%
 Time  (median):     357.156 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   373.151 μs ± 151.281 μs  ┊ GC (mean ± σ):  1.54% ±  3.54%

  ██▅▃▂▁▁▁          ▂▃▂▁                                        ▂
  ████████▇▆▆▆▅▆▅▅▆██████▇▆▆▅▅▄▄▃▅▅▄▃▃▅▃▄▃▄▁▁▁▁▄▄▃▄▁▄▁▁▄▁▁▁▁▃▆▇ █
  353 μs        Histogram: log(frequency) by time        591 μs <

  Memory estimate: 99.72 KiB, allocs estimate: 2988.
```

After type optimization (23/10/22):

``` julia
julia> using BenchmarkTools, ProfileView, Debugger, Infiltrator, Revise, AbstractTrees, Test
julia> using Norg

julia> using Norg
[ Info: Precompiling Norg [64228516-68de-4e23-8112-6940bc1970d1]

julia> s = """
       [Neorg] is a fancy organizational tool for everyone.

       I like shilling [Neorg]{https://github.com/nvim-neorg/neorg} sorry :(
       """
"[Neorg] is a fancy organizational tool for everyone.\n\nI like shilling [Neorg]{https://github.com/nvim-neorg/neorg} sorry :(\n"

julia> @benchmark parse(Norg.AST.NorgDocument, s)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  43.719 μs …  4.510 ms  ┊ GC (min … max): 0.00% … 97.40%
 Time  (median):     44.881 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   58.875 μs ± 77.725 μs  ┊ GC (mean ± σ):  2.42% ±  1.94%

  █▃▄▄▃▃▁               ▁▁▂▁                            ▁▃    ▁
  ████████▆▇▇▇▆▆▆▆▇▇▇███████▇▇▇▇▇▆▆▇▆▆▆▆▅▅▄▆▄▅▄▄▅▄▃▄▅▄▅▄██▇▇█ █
  43.7 μs      Histogram: log(frequency) by time       143 μs <

 Memory estimate: 21.34 KiB, allocs estimate: 210.
@end

After some more optimization :

@code julia
julia> @benchmark parse(Norg.AST.NorgDocument, s)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  43.526 μs …  12.982 ms  ┊ GC (min … max): 0.00% … 98.94%
 Time  (median):     44.967 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   58.194 μs ± 184.435 μs  ┊ GC (mean ± σ):  4.40% ±  1.40%

  █▂▅▅▄▄▃▁                                              ▃      ▁
  ██████████▇▇▇▇▇▆▇▅▅▆▅▆▅▅▆▅▅▅▅▅▄▄▅▅▅▅▄▄▄▃▅▄▅▅▅▅▄▄▃▅▄▄▄▄█▇████ █
  43.5 μs       Histogram: log(frequency) by time       148 μs <

  Memory estimate: 25.20 KiB, allocs estimate: 288.
```